fix password generation
the old version of `tr` (i.e. the one that comes with MacOS) was breaking when
fed a byte sequence from /dev/urandom (i.e. a byte sequence that isn't acii or
utf-8 encoded text), setting the locale to 'C' will allow it to handle this

see https://unix.stackexchange.com/questions/141420/tr-complains-of-illegal-byte-sequence