fix #7: domain name ip lookup validation fails
We'll remove the validation and put a message in place warning that the
DNS records must be configured.

Our check for successfully obtaining a certificate,

    if [[ $? -ne 0 ]]; then

was not catching all of the failures to obtain a certificat, and we were
still modifiying the nginx config without a valid ssl cert, causing
nginx to fail. Adding

    set -e

to the start of the command will cause us to bail out early, so that we
don't modify the nginx config for a site (and halt the nginx process) if
we are not successful in obtaining a certificate from letsencrypt

see also https://github.com/zgulde/tomcat-setup/issues/4