Fix passwordless log viewing and service restarting
Previously users were not able to view logs or restart a site's service
without a password due to an issue with the file name that was defined
to do this.

We'll create a separate file with the permissions for manipulating the
service for this site so that it is easier to clean up. The name of this
file must not contain '.'s though, so we'll generate a file named after
the domain with '-'s instead of '.'s

see /etc/sudoers.d/README and `man 5 sudoers`