s/wget/curl/
A goal of this project is to be relatively dependency free. While not every
linux distro has curl on it by default (some do have wget though), MacOS has
curl included by default, which is what the majority of our users are on.

We were only using it in one place, so it's pretty easy to re-write.

In addition, we'll add a check for curl and give a useful error message if we
can't find it.