Fix group permissions issues
tl;dr: allow group write permissions by default

We'll set the default umask to 0002 (i.e. make files and folders
group-writable by default, normally this is 0022, or group-readable, but
not group-writable).

This is set for interactive login sessions in
/etc/profile.d/group_umask.sh, and anywhere we have a non-interactive
login we specify the umask as part of the script (including, e.g.
`myserver run`).

see:

- https://unix.stackexchange.com/questions/4921/sh-startup-files-over-ssh/4953#4953
- man umask