Be more safe when adding groups for a new user
Which groups to add a new user to are determined by the config files
that exist in /etc/nginx/sites-available.

If you've added a custom site configuration (this might be necessary
for example, with an app that needs https and handles wildcard
subdomains), then this group detection could fail.

To fix this, we'll still use the same group detection, but we'll first
check to make sure we have a valid group.