Don't fail silently
Having `IF NOT EXISTS` in the database creation query causes the command
to silently not do anything if the database already exists. cods will
then think that we successfully created a database and add new generated
credentials that don't actually do anything.

This commit will make the failure explicit so that we don't add useless
credentials to the credentials file.

Fixes https://github.com/zgulde/cods/issues/17