Rename config file to cods-config
Apparently spring boot will look for a `config` _directory_ in the
directory in which it is starting up to load configuration files from.

Cods also utilizes a `config` _file_ at /srv/example.com/config to
figure out how to copy sensitive files on the server.

This hasn't been an issue when last tested with springboot 2.1.x, but as
of 2.4.x the application fails to startup with a message about "config"
not being a directory.

Based on very brief testing it seems like this is only an issue when the
spring security dependency is added (!?). Rather that run it down or
reqeuire users to configure their spring boot apps differently, we'll
rename the cods configuration file.